Moved test file into separate directory
